+++ library/Kathu/World/Field.hs
@@ -16,7 +16,7 @@ import qualified Data.Vector.Unboxed.Mutable as UMVec
 import           Linear.V2                   (V2(..))
 
 import           Kathu.Entity.System         (Tiles, fromTiles)
-import           Kathu.Util.Flow             (mapPair, mapSnd)
+import           Kathu.Util.Flow             (mapPair, mapFst, mapSnd)
 import           Kathu.Util.Polygon
 import           Kathu.World.Tile
 
@@ -153,15 +153,17 @@ fromTileVector2D fgTiles = if yLayers == 0 || xLayers == 0 then empty else build
           --debugPrint fs = (liftIO . putStrLn . concat $ ["x-len ", show xLen, "; y-len ", show yLen, "; x-layers ", show xLayers, "; y-layers ", show yLayers]) >> pure fs
 
 -- | Transforms a FieldSet into a list of V2 lists defining collision shapes
-mkCollisionPolygons :: MonadIO m => Tiles g -> FieldSet -> m [[V2 Double]]
-mkCollisionPolygons tiles = fmap (concat . fmap mkTriangles) . mapM isSolidVec . Map.assocs
+mkCollisionPolygons :: MonadIO m => Tiles g -> FieldSet -> m (Vec.Vector [V2 Double])
+mkCollisionPolygons tiles = fmap (Vec.concat . fmap mkTriangles) . mapM isSolidVec . Map.assocs
     where isSolidTS :: MonadIO m => TileState -> m Bool
           isSolidTS ts = view isSolid <$> (liftIO . fromTiles tiles $ ts)
           -- transforms field into a 1D vector of bools for if is solid; accessed at pos with indexFromCoord
           isSolidVec :: MonadIO m => (V2 Int, Field) -> m (V2 Int, UVec.Vector Bool)
           isSolidVec (pos, f) = fmap ((pos,) . UVec.fromList . reverse) . fieldFoldWithEmptyM (\acc _ ts -> (:acc) <$> isSolidTS ts) [] $ f
-          mkTriangles :: (V2 Int, UVec.Vector Bool) -> [[V2 Double]]
-          mkTriangles ((V2 wx wy), v) = uncurry (++)
-                                      . mapSnd (concat . fmap triangulate)
-                                      . mapPair (fmap (fmap ((+(V2 (-0.5) (-1))) . fmap fromIntegral . (+(V2 (wx * fieldDim) (wy * fieldDim))))))
+          
+          mkTriangles :: (V2 Int, UVec.Vector Bool) -> Vec.Vector [V2 Double]
+          mkTriangles ((V2 wx wy), v) = uncurry (Vec.++)
+                                      . mapFst  (Vec.map polyBorder)
+                                      . mapSnd  (Vec.concat . Vec.toList . Vec.map (Vec.fromList . triangulate))
+                                      . mapPair (Vec.map (mapPolyVertices ((+(V2 (-0.5) (-1))) . fmap fromIntegral . (+(V2 (wx * fieldDim) (wy * fieldDim))))))
                                       $ convexAndConcaveFromBinaryGrid v fieldDim fieldDim